+++ addons/confirma/src/extensions/ConfirmArrayExtensions.cs
@@ -1,7 +1,7 @@
 using System;
 using Confirma.Exceptions;
 
-namespace Confirma;
+namespace Confirma.Extensions;
 
 public static class ConfirmArrayExtensions
 {