+++ addons/confirma/src/extensions/ConfirmFileExtensions.cs
@@ -2,7 +2,7 @@ using System.IO;
 using Confirma.Exceptions;
 using Godot;
 
-namespace Confirma;
+namespace Confirma.Extensions;
 
 public static class ConfirmFileExtensions
 {