+++ tests/ConfirmEqualTest.cs
@@ -1,10 +1,11 @@
 using Confirma.Attributes;
 using Confirma.Exceptions;
+using Confirma.Extensions;
 
 namespace Confirma.Tests;
 
 [TestClass]
-[Parallelizable()]
+[Parallelizable]
 public static class ConfirmEqualTest
 {
 	[TestCase(1, 1)]