+++ tests/ConfirmTest.cs
@@ -1,11 +1,12 @@
 using System;
 using Confirma.Attributes;
 using Confirma.Exceptions;
+using Confirma.Extensions;
 
 namespace Confirma.Tests;
 
 [TestClass]
-[Parallelizable()]
+[Parallelizable]
 public static class ConfirmTest
 {
 	[TestCase]