+++ tests/ConfirmVectorTest.cs
@@ -1,6 +1,7 @@
 using Godot;
 using Confirma.Attributes;
 using Confirma.Exceptions;
+using Confirma.Extensions;
 
 namespace Confirma.Tests;